% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{Cleanup / delete opendata files}
\usage{
cleanup(name, dir = get_dir())
}
\arguments{
\item{name}{character, dataset name or NULL}

\item{dir}{character, opendata directory}
}
\value{
no return value
}
\description{
Cleanup / delete opendata files
}
\examples{
cleanup()
}
