% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_spec.R
\name{create_data_spec}
\alias{create_data_spec}
\title{Specify a new dataset}
\usage{
create_data_spec(name)
}
\arguments{
\item{name}{character, name of the new dataset}
}
\value{
no return value
}
\description{
Specify a new dataset
}
\details{
This function is not exported, but only used in development.
Use via opendata:::specify_data() or by utilizing devtools::load_all().
}
\examples{
# opendata:::create_data_spec("abc")
}
