% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_data.R
\name{data_fin_struktuuritoetus}
\alias{data_fin_struktuuritoetus}
\title{Struktuuritoetuste info}
\usage{
data_fin_struktuuritoetus()
}
\description{
Opendata.ee riigi struktuuritoetuste nimekiri
}
\details{
Stuktuuritoetused
}

