% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_data.R
\name{data_riigiteenused}
\alias{data_riigiteenused}
\title{Riigiteenuste registri info}
\usage{
data_riigiteenused()
}
\arguments{
\item{none}{}
}
\description{
Riigiteenuste andmebaas
}
\details{
Riigiteenuste registri info
}

