% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_wdbc.R
\name{data_wdbc}
\alias{data_wdbc}
\alias{get_info_wdbc}
\alias{proc_data_wdbc}
\alias{get_data_wdbc}
\title{Download, process and get information for the \strong{wdbc} dataset.}
\usage{
get_info_wdbc(...)

proc_data_wdbc(...)

get_data_wdbc(method = "curl", ...)
}
\arguments{
\item{...}{potential further arguments}

\item{method}{character, see \link{download.file}}
}
\value{
see details
}
\description{
Download, process and get information for the \strong{wdbc} dataset.
}
\details{
"Features are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass.
They describe characteristics of the cell nuclei present in the image. A few of the images can
be found at \url{http://www.cs.wisc.edu/~street/images/}.
Separating plane described above was obtained using Multisurface Method-Tree (MSM-T)\eqn{^2},
a classification method which uses linear programming to construct a decision tree.
Relevant features were selected using an exhaustive search in the space of 1-4 features
and 1-3 separating planes.
The actual linear program used to obtain the separating plane in the 3-dimensional space is
that described in: Bennett and Mangasarian (1992)\eqn{^3}." \eqn{^1}

\link{get_info}("wdbc"): Return a list with properties of the processed dataset
and metadata of the data source.

\link{load_data}("wdbc"): Load processed data into memory.
If the data has not been processed before, also runs proc_data("wdbc").

\link{proc_data}("wdbc"): Process the original data and save an rds file (for later use).
If the data has not been downloaded before, also runs get_data("wdbc").

\link{get_data}("wdbc"): Download the original data from the internet and save it locally.

The main reason the "_wdbc" functions are exported, is to allow easy inspection of the source code.
In practice, the use of the wrapper versions without suffix "_wdbc"
is recommended, as shown in the examples.
}
\examples{
get_info("wdbc") # get_info_wdbc()
get_data("wdbc") # get_data_wdbc()
proc_data("wdbc") # proc_data_wdbc()
load_data("wdbc") 

}
\references{
(1) \url{https://archive.ics.uci.edu/ml/datasets/breast+cancer+wisconsin+(diagnostic)}
(2) K. P. Bennett, "Decision Tree Construction Via Linear Programming." Proceedings of the 4th
Midwest Artificial Intelligence and Cognitive Science Society, pp. 97-101, 1992
(3) K. P. Bennett and O. L. Mangasarian: "Robust Linear Programming Discrimination of Two
Linearly Inseparable Sets", Optimization Methods and Software 1, 1992, 23-34.
}
\seealso{
\link{get_info}, \link{get_data}, \link{proc_data}, \link{load_data}
}
