% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{Description of the datasets covered by opendata}
\format{
A nested list whereby each (first level) element describes a dataset and has elements
\describe{
\item{properties}{describe the processed dataset, e.g. length, nrow, ncol, ...}
\item{metadata}{describe the original data (source), e.g. repository, url, ...}
}
}
\usage{
datasets
}
\description{
A dataset describing properties and metadata of all datasets that can be accessed with opendata.
}
\keyword{datasets}
