% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir.R
\name{delete_config_file}
\alias{delete_config_file}
\title{Delete the opendata config file}
\usage{
delete_config_file()
}
\value{
no return value
}
\description{
Delete the opendata config file
}
\examples{
delete_config_file()
}
