% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download an opendata dataset from the internet}
\usage{
get_data(name, ...)
}
\arguments{
\item{name}{character, name of the dataset}

\item{...}{further arguments (passed to get_data_xyz for name="xyz")}
}
\value{
name (to allow piping)
}
\description{
Download an opendata dataset from the internet
}
\examples{
get_data("ctg")
}
