% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir.R
\name{get_dir}
\alias{get_dir}
\title{Get the opendata directory for the current project}
\usage{
get_dir()
}
\value{
returns the data directory as a character, the default
is "./data" (if not set before via \link{set_dir})
}
\description{
Get the opendata directory for the current project
}
\examples{
get_dir()
}
