% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get information for a dataset}
\usage{
get_info(name)
}
\arguments{
\item{name}{character, name of the dataset}
}
\value{
a list of properties and metadata
}
\description{
Get information for a dataset
}
\details{
get_info("xyz") is analogous to get_info_xyz()
}
\examples{
get_info("ctg")
}
