% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitalrank_hcm_2017.R
\docType{data}
\name{hospitalrank_hcm_2017}
\alias{hospitalrank_hcm_2017}
\title{Rank of hospitals at Ho Chi Minh city in 2017}
\format{A data frame with 101 rows and 5 variables:
\describe{
  \item{rank}{Hospital rank (decided by the Health Department at Ho Chi Minh city)}
  \item{hospital}{Hospital name}
  \item{quality_score}{Mean quality score of the hospital}
  \item{satisfactory_score}{Mean patient's satisfactory score of the hospital}
  \item{type}{Type of hospital}
}}
\source{
\url{http://thanhnien.vn/suc-khoe/top-cac-benh-vien-co-diem-chat-luong-cao-nhat-tai-tphcm-790191.html}
}
\usage{
hospitalrank_hcm_2017
}
\description{
A dataset containing the mean quality score, patient's satisfactory score and the rank of 100 hospitals at Ho Chi Minh city in 2017
diamonds.
}
\keyword{datasets}
