% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load an opendata dataset}
\usage{
load_data(name, keep_files = TRUE, keep_rds = TRUE)
}
\arguments{
\item{name}{character}

\item{keep_files}{logical}

\item{keep_rds}{logical}
}
\value{
data in suitable format
}
\description{
Load an opendata dataset
}
\examples{
load_data("ctg")
}
