% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_data.R
\name{proc_data}
\alias{proc_data}
\title{Process an opendata dataset}
\usage{
proc_data(name, ..., keep_files = TRUE)
}
\arguments{
\item{name}{character, name of opendata dataset}

\item{...}{further arguments (passed to proc_data_xyz for name="xyz")}

\item{keep_files}{logical, should the original files be kept? (default: TRUE)}
}
\value{
no return value
}
\description{
Process an opendata dataset
}
