% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir.R
\name{set_dir}
\alias{set_dir}
\title{Set the opendata directory for the current project}
\usage{
set_dir(dir = "~/R/opendata")
}
\arguments{
\item{dir}{character, specification of directory, the default is "~/R/opendata"
(within the current working directory)}
}
\description{
Set the opendata directory for the current project
}
\details{
the directory is saved in the config.yml file created
in the current working directory
}
\examples{
# get_dir()
# set_dir("~/R/opendata")
# get_dir()
}
